Flyway Database Migrations added
Two migration files have been added.
V1 - to create initial tables
V2- add foreign keys

+++ backend_PharmaPulse/src/main/java/com/group03/backend_PharmaPulse/purchase/internal/entity/PurchaseGroup.java
@@ -16,7 +16,7 @@ public class PurchaseGroup {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "purchase_group_id",length = 50)
-    private int purchaseGroupId;
+    private Long purchaseGroupId;
 
     private String purchaseGroupName;
